* Changed from char array to stringstream in several areas in the code for security reasons. * Rewrote guild commands code, not so deep code nesting anymore and a lot of bugs were fixed. * Possibility to disable rule violation reports code again, due to possible crashes. * /info command will now display account warnings. * Extended /newtype looktype range for new 8.5 looktypes. * New effects for /ghost command. * Fixed a major bug with multiclient check command. * Query destination improvements, will now access deeper containers. * Server save code improvements. * Performance improvements to queries which only fetches a single row (this only applies to queries where the filter isn't unique). * Optimized online status updating code. * Free premium code fixes. * Improved error messages during map load, most of them will now print the position. * New Lua functions: isPlayerPzLocked(cid), isDepot(uid) * Code cleanup and fixes to Lua functions: doPlayerAddPremiumDays and doPlayerRemovePremiumDays. * Fixed a bug with monsters which would cause summons to freeze if a player changed floor during its death delay. * NPC walkinterval value will no longer be overwritten by deprecated autowalk. * Players with access level will now be able to wear all outfits with all available addons. * Excessive Unjustified Killing doesn't give any account warning, so it shouldn't be giving account deletion either if player has enough warnings for that. * Protocol 8.5 support: 	New icons: pigeon (protection zone) & red swords (protection zone-lock). 	Loot message. 	Dealt damage message. 	New looktypes (including wedding outfits). 	New objects. 	New doors & windows.
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/branches/0.2@2342 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ map.h
@@ -186,7 +186,7 @@ class Map
 		  * Get a single tile.
 		  * \returns A pointer to that tile.
 		  */
-		Tile* getTile(uint16_t x, uint16_t y, uint16_t z);
+		Tile* getTile(int32_t x, int32_t y, int32_t z);
 		Tile* getTile(const Position& pos);
 
 		uint32_t clean();
